Refactor(wiring): move committer out of srv deps and rename it to client deps. Still awkward placement for commmitter.
