+++ wiring/module_test.go
